Colorbars: Force cref closer to center of colorbar
Modify rounding of cref for non-log plots.
Before it could happen in edge cases that cref lies at 99% of
colorbar range.
--> This would cause loss of color information

Example case: salinity between 29.0 psu and 39.2 psu
--> cref should be at 34.1 psu (middle).
Without changes it is rounded to 30 (loop exits in first iteration)
With changes it is now forced to round to 34 --> closer to middle